+++ src/dawn/net/BitStream.h
@@ -11,33 +11,33 @@ namespace dw {
 // Note: Data must outlive the InputBitStream.
 class DW_API InputBitStream : public InputStream {
 public:
-    InputBitStream(const byte* data, u32 length);
+    InputBitStream(const byte* data, usize length);
     InputBitStream(const Vector<byte>& data);
     ~InputBitStream() = default;
 
     // InputStream.
-    u32 readData(void* dest, u32 size) override;
-    void seek(u64 position) override;
+    usize readData(void* dest, usize size) override;
+    void seek(usize position) override;
 
     const byte* data() const;
-    u32 length() const;
+    usize length() const;
 
 private:
     const byte* data_;
-    u32 length_;
+    usize length_;
 };
 
 class DW_API OutputBitStream : public OutputStream {
 public:
-    OutputBitStream(u32 bytes_to_reserve = 0);
+    OutputBitStream(usize bytes_to_reserve = 0);
     ~OutputBitStream() = default;
 
-    u32 writeData(const void* src, u32 size) override;
+    usize writeData(const void* src, usize size) override;
 
     const Vector<byte>& vec_data() const;
 
     const byte* data() const;
-    u32 length() const;
+    usize length() const;
 
 private:
     Vector<byte> data_;
