+++ src/dawn/renderer/Program.cpp
@@ -33,7 +33,7 @@ rhi::ProgramHandle Program::internalHandle() const {
 
 void Program::applyRendererState() {
     // Set textures.
-    for (size_t i = 0; i < texture_units_.size(); i++) {
+    for (uint i = 0; i < static_cast<uint>(texture_units_.size()); i++) {
         if (!texture_units_[i]) {
             break;
         }
